+++ SOD.Common/Helpers/PluginDetection.cs
@@ -1,9 +1,7 @@
-using System;
-using System.Linq;
 using BepInEx;
 using BepInEx.Configuration;
 using BepInEx.Unity.IL2CPP;
-using Il2CppSystem.Runtime.InteropServices;
+using System;
 using UniverseLib;
 
 namespace SOD.Common.Helpers
@@ -35,39 +33,12 @@ namespace SOD.Common.Helpers
         }
 
         /// <summary>
-        /// Get the full GUID of a plugin by searching for a partial GUID
-        /// across all loaded plugins.
-        /// </summary>
-        /// <remarks>
-        /// Should be called in response to <c>OnAllPluginsFinishedLoading</c>.
-        /// Useful for plugins such as Babbler, which use a prefix on their
-        /// GUID to influence load order.
-        /// </remarks>
-        /// <param name="partialPluginGuid">A partial GUID.</param>
-        /// <returns></returns>
-        /// <exception cref="System.InvalidOperationException"></exception>
-        public string GetPluginGuidFromPartialGuid(string partialPluginGuid)
-        {
-            var guids = IL2CPPChainloader.Instance.Plugins.Keys;
-            var matches = guids.Where(guid => guid.ToLower().Contains(partialPluginGuid.ToLower())).ToArray();
-            int matchCount = matches.Length;
-            if (matchCount == 0)
-            {
-                return null;
-            }
-            if (matchCount > 1)
-            {
-                throw new System.InvalidOperationException($"Multiple GUIDs match the provided partial GUID ({partialPluginGuid}): {string.Join(", ", matches)}.");
-            }
-            return matches.First();
-        }
-
-        /// <summary>
         /// Check if a BepInEx plugin is loaded given the plugin GUID.
         /// </summary>
         /// <param name="pluginGuid">The GUID of the target plugin.</param>
         /// <returns>True if the plugin has been loaded, false otherwise.</returns>
-        public bool IsPluginLoaded(string pluginGuid) => IL2CPPChainloader.Instance.Plugins.ContainsKey(pluginGuid);
+        public bool IsPluginLoaded(string pluginGuid) 
+            => IL2CPPChainloader.Instance.Plugins.ContainsKey(pluginGuid);
 
         /// <summary>
         /// Gets the plugin info (incompatibilities, version, user-friendly
@@ -80,7 +51,8 @@ namespace SOD.Common.Helpers
         /// <param name="info">The BepInEx PluginInfo data object of the target
         /// plugin.</param>
         /// <returns></returns>
-        public bool TryGetPluginInfo(string pluginGuid, out PluginInfo info) => IL2CPPChainloader.Instance.Plugins.TryGetValue(pluginGuid, out info);
+        public bool TryGetPluginInfo(string pluginGuid, out PluginInfo info) 
+            => IL2CPPChainloader.Instance.Plugins.TryGetValue(pluginGuid, out info);
 
         /// <summary>
         /// Gets the plugin info (incompatibilities, version, user-friendly
@@ -95,9 +67,7 @@ namespace SOD.Common.Helpers
         public PluginInfo GetPluginInfo(string pluginGuid)
         {
             if (!TryGetPluginInfo(pluginGuid, out var info))
-            {
                 throw new System.Collections.Generic.KeyNotFoundException($"Plugin GUID ({pluginGuid}) not found.");
-            }
             return info;
         }
 
@@ -120,9 +90,7 @@ namespace SOD.Common.Helpers
             var info = GetPluginInfo(pluginGuid);
             var plugin = (BasePlugin)info.Instance.TryCast(typeof(BasePlugin));
             if (!plugin.Config.TryGetEntry<T>(section: section, key: key, entry: out var entry))
-            {
-                throw new System.InvalidOperationException($"No configuration entry found for ({section}, {key}).");
-            }
+                throw new InvalidOperationException($"No configuration entry found for ({section}, {key}).");
             return entry.Value;
         }