+++ SOD.Common/ModFolderContent/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added extra AddHours method to TimeData struct in Lib.Time
 - Fixed parsing issue with GetFnvHashCode extension
 - Fixed Lib.Time not initializing properly in some specific scenarios
+- Removed 'GetPluginGuidFromPartialGuid' method from PluginDetection lib
 
 **1.1.8**
 - Added non-generic bind extension on BepInEx ConfigFile object